pyproject.toml
[project]
name = "auto-installer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.30.0",
    "pillow>=11.0.0",
    "pyautogui>=0.9.54",
]
